Refactor migration v3 to import unchanged models
