+++ migration/v2/model/base.go
@@ -4,6 +4,7 @@ import (
 	"time"
 )
 
+//
 // Model Base model.
 type Model struct {
 	ID         uint      `gorm:"<-:create;primaryKey"`
